requirements.txt
flask==3.1.0
flask-sqlalchemy==3.1.1
flask-migrate==4.1.0
flask-jwt-extended==4.7.1
flask-mail==0.10.0
flask-cors==5.0.1
flasgger==0.9.7.1
psycopg2-binary==2.9.10
python-dotenv==1.1.0
